move js code to a appropriate file
